refactor: split account_detail_screen.dart into 3 themed extensions
_AccountDetailScreenState was a 988-LoC State class. Extract three
contiguous method blocks into part files under lib/ui/screens/
account_detail/, each an extension on _AccountDetailScreenState:

- list_builders.dart: _buildEntries + _buildTxTile + _openTransaction
  + _addTransaction (~211 LoC)
- transaction_actions.dart: _flagAsIncome + _flagAsAdjustment +
  _confirmWipeTransactions + _confirmDeleteAccount + _editAccount
  (~238 LoC)
- balance_dialog.dart: _showBalanceDialog + _executeBalanceRecalc
  (~224 LoC)

account_detail_screen.dart: 1046 → 386 LoC (−63%). Build method,
lifecycle (initState/dispose), and the two setState callers stay
in the class body — no analyzer warnings.

Distribution across lib/:
  files > 1000 LoC: 4 → 3
  max:              1255 → 1225
  avg:              312 → 305

All four suites green: analyze clean, 778 unit, integration +4, live
data fetch +1.
