+++ lib/ui/screens/account_detail_screen.dart
@@ -22,6 +22,9 @@ import '../widgets/selection/selection_controller.dart';
 
 part 'account_detail/entries.dart';
 part 'account_detail/list_widgets.dart';
+part 'account_detail/list_builders.dart';
+part 'account_detail/transaction_actions.dart';
+part 'account_detail/balance_dialog.dart';
 
 final _log = getLogger('AccountDetailScreen');
 
@@ -376,669 +379,6 @@ class _AccountDetailScreenState extends ConsumerState<AccountDetailScreen> {
   /// display entries where each transfer collapses its two legs into a single
   /// [_TransferEntry]. Non-transfer txs become [_TxEntry]s. Original order is
   /// preserved; transfers slot in at the position of whichever leg came first.
-  List<_Entry> _buildEntries({
-    required List<Transaction> filtered,
-    required bool detectTransfers,
-    required int Function(DateTime) dayKey,
-  }) {
-    if (!detectTransfers || filtered.length < 2) {
-      return [for (final t in filtered) _TxEntry(t)];
-    }
-
-    final txById = <int, Transaction>{for (final t in filtered) t.id: t};
-    final transferOfTx = <int, _TransferEntry>{};
-
-    // Bucket by (day, currency, |amount in cents|).
-    final buckets = <String, ({List<Transaction> pos, List<Transaction> neg})>{};
-    for (final t in filtered) {
-      if (t.amount == 0) continue;
-      final cents = (t.amount.abs() * 100).round();
-      final k = '${dayKey(t.valueDate)}|${t.currency}|$cents';
-      final bucket = buckets[k] ?? (pos: <Transaction>[], neg: <Transaction>[]);
-      if (t.amount > 0) {
-        bucket.pos.add(t);
-      } else {
-        bucket.neg.add(t);
-      }
-      buckets[k] = bucket;
-    }
-    for (final bucket in buckets.values) {
-      final pos = bucket.pos;
-      final neg = bucket.neg;
-      if (pos.isEmpty || neg.isEmpty) continue;
-      // Greedy pair: each positive consumes the first available negative
-      // belonging to a different account.
-      final consumedNeg = <int>{};
-      for (final p in pos) {
-        Transaction? match;
-        for (final n in neg) {
-          if (consumedNeg.contains(n.id)) continue;
-          if (n.accountId == p.accountId) continue;
-          match = n;
-          break;
-        }
-        if (match == null) continue;
-        consumedNeg.add(match.id);
-        final pair = _TransferEntry(inflow: p, outflow: match);
-        transferOfTx[p.id] = pair;
-        transferOfTx[match.id] = pair;
-      }
-    }
-
-    final result = <_Entry>[];
-    final emitted = <_TransferEntry>{};
-    for (final t in filtered) {
-      final pair = transferOfTx[t.id];
-      if (pair != null) {
-        if (emitted.add(pair)) result.add(pair);
-      } else {
-        result.add(_TxEntry(t));
-      }
-    }
-    // Silence the unused warning for txById in release builds.
-    assert(txById.isNotEmpty);
-    return result;
-  }
-
-  Widget _buildTxTile({
-    required BuildContext context,
-    required Transaction tx,
-    required bool isPositive,
-    required dynamic rowAmtFmt,
-    required dynamic dateFmt,
-    required Map<int, String> accountNameById,
-    required AppStrings s,
-  }) {
-    final tile = ListTile(
-      dense: true,
-      leading: CircleAvatar(
-        radius: 16,
-        backgroundColor: isPositive
-            ? Colors.green.withValues(alpha: 0.1)
-            : Colors.red.withValues(alpha: 0.1),
-        child: Icon(
-          isPositive ? Icons.arrow_downward : Icons.arrow_upward,
-          size: 16,
-          color: isPositive ? Colors.green : Colors.red,
-        ),
-      ),
-      title: Text(
-        tx.description.isNotEmpty ? tx.description : s.noDescription,
-        maxLines: 1,
-        overflow: TextOverflow.ellipsis,
-        style: const TextStyle(fontSize: 14),
-      ),
-      subtitle: _isReadOnly
-          ? Row(
-              children: [
-                Text(dateFmt.format(tx.valueDate),
-                    style: const TextStyle(fontSize: 12)),
-                const SizedBox(width: 6),
-                Flexible(
-                  child: Container(
-                    padding: const EdgeInsets.symmetric(
-                        horizontal: 6, vertical: 1),
-                    decoration: BoxDecoration(
-                      color: Theme.of(context)
-                          .colorScheme
-                          .secondaryContainer,
-                      borderRadius: BorderRadius.circular(4),
-                    ),
-                    child: Text(
-                      accountNameById[tx.accountId] ?? '#${tx.accountId}',
-                      style: TextStyle(
-                        fontSize: 11,
-                        color: Theme.of(context)
-                            .colorScheme
-                            .onSecondaryContainer,
-                      ),
-                      maxLines: 1,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                  ),
-                ),
-              ],
-            )
-          : Text(dateFmt.format(tx.valueDate),
-              style: const TextStyle(fontSize: 12)),
-      trailing: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          PrivacyText(
-            '${isPositive ? '+' : ''}${rowAmtFmt.format(tx.amount)}',
-            style: TextStyle(
-              fontWeight: FontWeight.bold,
-              color: isPositive ? Colors.green.shade700 : Colors.red.shade700,
-              fontSize: 14,
-            ),
-          ),
-          if (!_isReadOnly) ...[
-            const SizedBox(width: 4),
-            PopupMenuButton<String>(
-              iconSize: 18,
-              padding: EdgeInsets.zero,
-              tooltip: isPositive
-                  ? s.flagAsIncomeTooltip
-                  : s.flagAsAdjustmentTooltip,
-              itemBuilder: (_) => [
-                if (isPositive)
-                  PopupMenuItem(
-                    value: 'flag_income',
-                    child: Row(
-                      children: [
-                        const Icon(Icons.label_important_outline, size: 18),
-                        const SizedBox(width: 8),
-                        Text(s.flagAsIncomeTooltip),
-                      ],
-                    ),
-                  )
-                else
-                  PopupMenuItem(
-                    value: 'flag_adjustment',
-                    child: Row(
-                      children: [
-                        const Icon(Icons.compare_arrows, size: 18),
-                        const SizedBox(width: 8),
-                        Text(s.flagAsAdjustmentTooltip),
-                      ],
-                    ),
-                  ),
-              ],
-              onSelected: (v) {
-                if (v == 'flag_income') _flagAsIncome(tx);
-                if (v == 'flag_adjustment') _flagAsAdjustment(tx);
-              },
-            ),
-          ],
-        ],
-      ),
-      onTap: _isReadOnly ? null : () => _openTransaction(tx),
-    );
-    return _isReadOnly
-        ? tile
-        : SelectableItem<int>(
-            controller: _selection,
-            id: tx.id,
-            child: tile,
-          );
-  }
-
-  void _openTransaction(Transaction tx) {
-    Navigator.push(
-      context,
-      MaterialPageRoute(
-        builder: (_) => TransactionEditScreen(
-          transaction: tx,
-          account: widget.account,
-        ),
-      ),
-    );
-  }
-
-  void _addTransaction() {
-    Navigator.push(
-      context,
-      MaterialPageRoute(
-        builder: (_) => TransactionEditScreen(account: widget.account),
-      ),
-    );
-  }
-
-  Future<void> _flagAsIncome(Transaction tx) async {
-    final s = ref.read(appStringsProvider);
-    var selectedType = IncomeType.income;
-
-    String typeLabel(IncomeType t) => switch (t) {
-      IncomeType.income              => s.incomeTypeIncome,
-      IncomeType.refund              => s.incomeTypeRefund,
-      IncomeType.pensionContribution => s.incomeTypePensionContribution,
-    };
-
-    final confirmed = await showDialog<bool>(
-      context: context,
-      builder: (ctx) => StatefulBuilder(
-        builder: (ctx, setDialogState) => AlertDialog(
-          title: Text(s.flagAsIncomeTitle),
-          content: DropdownButtonFormField<IncomeType>(
-            initialValue: selectedType,
-            decoration: InputDecoration(labelText: s.incomeTypeLabel),
-            items: IncomeType.values
-                .map((t) => DropdownMenuItem(value: t, child: Text(typeLabel(t))))
-                .toList(),
-            onChanged: (v) => setDialogState(() => selectedType = v!),
-          ),
-          actions: [
-            TextButton(onPressed: () => Navigator.pop(ctx, false), child: Text(s.cancel)),
-            FilledButton(onPressed: () => Navigator.pop(ctx, true), child: Text(s.add)),
-          ],
-        ),
-      ),
-    );
-
-    if (confirmed != true) return;
-
-    await ref.read(incomeServiceProvider).create(
-      date: tx.valueDate,
-      amount: tx.amount,
-      type: selectedType,
-      currency: tx.currency,
-    );
-
-    if (mounted) {
-      showInfoSnack(context, s.incomeFlaggedSnack);
-    }
-  }
-
-  /// Mark a NEGATIVE transaction as a manual adjustment against an Inflow
-  /// `ExtraordinaryEvent`. The user picks the inflow record (best-match
-  /// proposed by smallest |valueDate − eventDate|, currency-equal first).
-  /// `addManualEntry` flips the sign for inflow events, so the resulting
-  /// entry on the inflow has `+|tx.amount|`.
-  Future<void> _flagAsAdjustment(Transaction tx) async {
-    final s = ref.read(appStringsProvider);
-    final locale = ref.read(appLocaleProvider).value ?? Platform.localeName;
-    final amtFmt = fmt.currencyFormat(locale, tx.currency);
-
-    final allEvents = await ref.read(extraordinaryEventsProvider.future);
-    final inflows = allEvents
-        .where((e) => e.direction == EventDirection.inflow && e.isActive)
-        .toList();
-    if (inflows.isEmpty) {
-      if (mounted) showInfoSnack(context, s.noInflowEventsAvailable);
-      return;
-    }
-
-    int bestMatchScore(ExtraordinaryEvent e) {
-      // Lower is better. Currency mismatch adds a huge penalty so same-currency
-      // candidates always win when present.
-      final dayDelta =
-          (e.eventDate.difference(tx.valueDate).inDays).abs();
-      final currencyPenalty = e.currency == tx.currency ? 0 : 1000000;
-      return currencyPenalty + dayDelta;
-    }
-
-    inflows.sort((a, b) => bestMatchScore(a).compareTo(bestMatchScore(b)));
-    var selectedId = inflows.first.id;
-    final dateFmt = fmt.shortDateFormat(locale);
-
-    if (!mounted) return;
-    final confirmed = await showDialog<bool>(
-      context: context,
-      builder: (ctx) => StatefulBuilder(
-        builder: (ctx, setDialogState) => AlertDialog(
-          title: Text(s.flagAsAdjustmentTitle),
-          content: Column(
-            mainAxisSize: MainAxisSize.min,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(s.flagAsAdjustmentBody(amtFmt.format(tx.amount.abs()))),
-              const SizedBox(height: 12),
-              DropdownButtonFormField<int>(
-                initialValue: selectedId,
-                isExpanded: true,
-                decoration: InputDecoration(labelText: s.flagAsAdjustmentInflow),
-                items: inflows
-                    .map((e) => DropdownMenuItem(
-                          value: e.id,
-                          child: Text(
-                            '${e.name} · ${e.currency} · ${dateFmt.format(e.eventDate)}',
-                            overflow: TextOverflow.ellipsis,
-                          ),
-                        ))
-                    .toList(),
-                onChanged: (v) =>
-                    setDialogState(() => selectedId = v ?? selectedId),
-              ),
-            ],
-          ),
-          actions: [
-            TextButton(onPressed: () => Navigator.pop(ctx, false), child: Text(s.cancel)),
-            FilledButton(onPressed: () => Navigator.pop(ctx, true), child: Text(s.add)),
-          ],
-        ),
-      ),
-    );
-
-    if (confirmed != true) return;
-
-    await ref.read(extraordinaryEventServiceProvider).addManualEntry(
-          eventId: selectedId,
-          date: tx.valueDate,
-          amount: tx.amount.abs(),
-          description: tx.description,
-        );
-
-    if (mounted) {
-      showInfoSnack(context, s.adjustmentFlaggedSnack);
-    }
-  }
-
-  Future<void> _confirmWipeTransactions(BuildContext context) async {
-    final s = ref.read(appStringsProvider);
-    final txCount = ref.read(accountTransactionsProvider(widget.account.id)).value?.length ?? 0;
-    if (txCount == 0) {
-      showInfoSnack(context, s.noTransactionsToWipe);
-      return;
-    }
-    final confirmed = await showConfirmDialog(
-      context,
-      title: s.wipeAllTransactionsTitle,
-      content: '${s.wipeTransactionsBody(widget.account.name)}${s.cannotBeUndone}',
-      confirmLabel: s.wipe,
-      cancelLabel: s.cancel,
-      confirmColor: Colors.orange,
-    );
-    if (confirmed) {
-      _log.warning('wiping transactions for account ${widget.account.id}');
-      final deleted = await ref.read(transactionServiceProvider).deleteByAccount(widget.account.id);
-      if (context.mounted) {
-        showInfoSnack(context, s.wipedTransactions(deleted));
-      }
-    }
-  }
-
-  Future<void> _confirmDeleteAccount(BuildContext context) async {
-    final s = ref.read(appStringsProvider);
-    final confirmed = await showConfirmDialog(
-      context,
-      title: s.deleteAccountTitle,
-      content: s.deleteAccountConfirm(widget.account.name),
-      confirmLabel: s.delete,
-      cancelLabel: s.cancel,
-      confirmColor: Colors.red,
-    );
-    if (confirmed) {
-      _log.warning('deleting account id=${widget.account.id} name=${widget.account.name}');
-      await ref.read(accountServiceProvider).delete(widget.account.id);
-      if (context.mounted) Navigator.pop(context);
-    }
-  }
-
-  Future<void> _editAccount(BuildContext context) async {
-    final s = ref.read(appStringsProvider);
-    final nameCtrl = TextEditingController(text: widget.account.name);
-    var currencyCtrl = TextEditingController(text: widget.account.currency);
-    var institutionCtrl = TextEditingController(text: widget.account.institution);
-    var isActive = widget.account.isActive;
-
-    await showDialog(
-      context: context,
-      builder: (ctx) => StatefulBuilder(
-        builder: (ctx, setDialogState) => AlertDialog(
-          title: Text(s.editAccountTitle),
-          content: SingleChildScrollView(
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                TextField(
-                  controller: nameCtrl,
-                  decoration: InputDecoration(labelText: s.name),
-                ),
-                const SizedBox(height: 8),
-                TextField(
-                  controller: currencyCtrl,
-                  decoration: InputDecoration(labelText: s.currency, hintText: 'EUR'),
-                ),
-                const SizedBox(height: 8),
-                TextField(
-                  controller: institutionCtrl,
-                  decoration: InputDecoration(labelText: s.institution),
-                ),
-                const SizedBox(height: 8),
-                SwitchListTile(
-                  title: Text(s.active),
-                  value: isActive,
-                  onChanged: (v) => setDialogState(() => isActive = v),
-                  contentPadding: EdgeInsets.zero,
-                ),
-              ],
-            ),
-          ),
-          actions: [
-            TextButton(onPressed: () => Navigator.pop(ctx), child: Text(s.cancel)),
-            FilledButton(
-              onPressed: () async {
-                if (nameCtrl.text.trim().isEmpty) return;
-                await ref.read(accountServiceProvider).update(
-                  widget.account.id,
-                  AccountsCompanion(
-                    name: Value(nameCtrl.text.trim()),
-                    currency: Value(currencyCtrl.text.trim()),
-                    institution: Value(institutionCtrl.text.trim()),
-                    isActive: Value(isActive),
-                    updatedAt: Value(DateTime.now()),
-                  ),
-                );
-                if (ctx.mounted) Navigator.pop(ctx);
-              },
-              child: Text(s.save),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Future<void> _showBalanceDialog(BuildContext context) async {
-    final s = ref.read(appStringsProvider);
-    // Get all transactions with rawMetadata to discover available columns
-    final txs = await ref.read(transactionServiceProvider).getByAccount(widget.account.id);
-    if (txs.isEmpty) {
-      if (context.mounted) {
-        showInfoSnack(context, s.noTransactionsToRecalc);
-      }
-      return;
-    }
-
-    // Discover columns from rawMetadata
-    final allColumns = <String>{};
-    for (final tx in txs) {
-      if (tx.rawMetadata != null) {
-        final meta = jsonDecode(tx.rawMetadata!) as Map<String, dynamic>;
-        allColumns.addAll(meta.keys);
-      }
-    }
-    final columns = allColumns.toList()..sort();
-
-    // Load saved config for current balance mode
-    final savedConfig = await ref.read(importConfigServiceProvider).getByAccount(widget.account.id);
-    Map<String, dynamic> savedMappings = {};
-    if (savedConfig != null) {
-      savedMappings = jsonDecode(savedConfig.mappingsJson) as Map<String, dynamic>;
-    }
-
-    var balanceMode = (savedMappings['__balanceMode'] as String?) ?? 'cumulative';
-    String? filterColumn = savedMappings['__balanceFilterColumn'] as String?;
-    if (filterColumn != null && !columns.contains(filterColumn)) filterColumn = null;
-    final filterInclude = <String>{};
-    if (savedMappings.containsKey('__balanceFilterInclude')) {
-      filterInclude.addAll(
-        (jsonDecode(savedMappings['__balanceFilterInclude'] as String) as List<dynamic>).cast<String>(),
-      );
-    }
-    // Get unique values for filter column from rawMetadata
-    List<String> uniqueValues(String col) {
-      final vals = <String>{};
-      for (final tx in txs) {
-        if (tx.rawMetadata == null) continue;
-        final meta = jsonDecode(tx.rawMetadata!) as Map<String, dynamic>;
-        final v = (meta[col]?.toString() ?? '').trim();
-        if (v.isNotEmpty) vals.add(v);
-      }
-      return vals.toList()..sort();
-    }
-
-    if (!context.mounted) return;
-
-    await showDialog(
-      context: context,
-      builder: (ctx) => StatefulBuilder(
-        builder: (ctx, setDialogState) => AlertDialog(
-          title: Text(s.recalcBalanceTitle),
-          content: ConstrainedBox(
-            constraints: const BoxConstraints(maxWidth: 500),
-            child: SingleChildScrollView(
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(s.recalcBalanceHelp,
-                      style: const TextStyle(fontSize: 13, color: Colors.grey)),
-                  const SizedBox(height: 12),
-                  SegmentedButton<String>(
-                    segments: [
-                      ButtonSegment(value: 'cumulative', label: Text(s.recalcCumulative)),
-                      ButtonSegment(value: 'column', label: Text(s.recalcColumn)),
-                      ButtonSegment(value: 'filtered', label: Text(s.recalcFiltered)),
-                    ],
-                    selected: {balanceMode},
-                    onSelectionChanged: (v) => setDialogState(() {
-                      balanceMode = v.first;
-                      if (balanceMode != 'filtered') {
-                        filterColumn = null;
-                        filterInclude.clear();
-                      }
-                    }),
-                  ),
-                  const SizedBox(height: 12),
-
-                  if (balanceMode == 'column')
-                    Text(
-                      s.balanceFromColumnHelp,
-                      style: const TextStyle(fontSize: 12, color: Colors.grey, fontStyle: FontStyle.italic),
-                    ),
-
-                  if (balanceMode == 'cumulative')
-                    Text(
-                      s.balanceCumulativeHelp,
-                      style: const TextStyle(fontSize: 12, color: Colors.grey, fontStyle: FontStyle.italic),
-                    ),
-
-                  if (balanceMode == 'filtered') ...[
-                    Text(s.filterColumnLabel, style: const TextStyle(fontWeight: FontWeight.w500, fontSize: 13)),
-                    const SizedBox(height: 4),
-                    DropdownButtonFormField<String>(
-                      initialValue: filterColumn,
-                      isExpanded: true,
-                      decoration: const InputDecoration(
-                        isDense: true,
-                        contentPadding: EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-                        border: OutlineInputBorder(),
-                      ),
-                      items: [
-                        DropdownMenuItem(value: null, child: Text('\u2014 ${s.none} \u2014', style: const TextStyle(color: Colors.grey))),
-                        ...columns.map((c) => DropdownMenuItem(value: c, child: Text(c))),
-                      ],
-                      onChanged: (v) => setDialogState(() {
-                        filterColumn = v;
-                        filterInclude.clear();
-                        if (v != null) filterInclude.addAll(uniqueValues(v));
-                      }),
-                    ),
-                    if (filterColumn != null) ...[
-                      const SizedBox(height: 8),
-                      Row(
-                        children: [
-                          Text(s.includeValues, style: const TextStyle(fontSize: 12, fontWeight: FontWeight.w500)),
-                          const Spacer(),
-                          TextButton(
-                            onPressed: () => setDialogState(() => filterInclude.addAll(uniqueValues(filterColumn!))),
-                            style: TextButton.styleFrom(visualDensity: VisualDensity.compact),
-                            child: Text(s.all, style: const TextStyle(fontSize: 11)),
-                          ),
-                          TextButton(
-                            onPressed: () => setDialogState(() => filterInclude.clear()),
-                            style: TextButton.styleFrom(visualDensity: VisualDensity.compact),
-                            child: Text(s.none, style: const TextStyle(fontSize: 11)),
-                          ),
-                        ],
-                      ),
-                      Wrap(
-                        spacing: 4,
-                        runSpacing: 0,
-                        children: uniqueValues(filterColumn!).map((val) {
-                          final selected = filterInclude.contains(val);
-                          return FilterChip(
-                            label: Text(val, style: const TextStyle(fontSize: 12)),
-                            selected: selected,
-                            onSelected: (v) => setDialogState(() {
-                              if (v) { filterInclude.add(val); } else { filterInclude.remove(val); }
-                            }),
-                          );
-                        }).toList(),
-                      ),
-                    ],
-                  ],
-
-
-
-                ],
-              ),
-            ),
-          ),
-          actions: [
-            TextButton(onPressed: () => Navigator.pop(ctx), child: Text(s.cancel)),
-            FilledButton(
-              onPressed: (balanceMode == 'filtered' && filterColumn == null) ||
-                      (balanceMode == 'column' && savedMappings['balanceAfter'] == null)
-                  ? null
-                  : () async {
-                      Navigator.pop(ctx);
-                      await _executeBalanceRecalc(txs, balanceMode, filterColumn, filterInclude, savedMappings);
-                      // Update saved config with new balance mode
-                      final updatedMappings = Map<String, dynamic>.from(savedMappings);
-                      updatedMappings['__balanceMode'] = balanceMode;
-                      if (filterColumn != null) {
-                        updatedMappings['__balanceFilterColumn'] = filterColumn;
-                      } else {
-                        updatedMappings.remove('__balanceFilterColumn');
-                      }
-                      if (filterInclude.isNotEmpty) {
-                        updatedMappings['__balanceFilterInclude'] = jsonEncode(filterInclude.toList());
-                      } else {
-                        updatedMappings.remove('__balanceFilterInclude');
-                      }
-                      await ref.read(importConfigServiceProvider).save(
-                        accountId: widget.account.id,
-                        skipRows: savedConfig?.skipRows ?? 0,
-                        mappings: updatedMappings.map((k, v) => MapEntry(k, v as String?)),
-                        formula: savedConfig != null
-                            ? (jsonDecode(savedConfig.formulaJson) as List<dynamic>)
-                                .map((e) => (e as Map<String, dynamic>).map((k, v) => MapEntry(k, v as String)))
-                                .toList()
-                            : [],
-                        hashColumns: savedConfig != null
-                            ? (jsonDecode(savedConfig.hashColumnsJson) as List<dynamic>).cast<String>()
-                            : [],
-                      );
-                    },
-              child: Text(s.recalculate),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Future<void> _executeBalanceRecalc(
-    List<Transaction> transactions,
-    String balanceMode,
-    String? filterColumn,
-    Set<String> filterInclude,
-    Map<String, dynamic> mappings,
-  ) async {
-    _log.info('balanceRecalc: mode=$balanceMode, filterCol=$filterColumn, include=$filterInclude, ${transactions.length} txs');
-    final s = ref.read(appStringsProvider);
-    final txSvc = ref.read(transactionServiceProvider);
-    final updated = await txSvc.recalculateBalances(
-      widget.account.id,
-      balanceMode: balanceMode,
-      savedMappings: mappings,
-    );
-    if (mounted) {
-      showInfoSnack(context, s.recalculatedBalances(updated));
-    }
-  }
 }
 
 /// A row in the All-accounts list. Either a single transaction or an
